chore: move bussiness logic to model
